+++ src/microservices/Microservices.CohortExtractor/Execution/RequestFulfillers/IExtractionRequestFulfiller.cs
@@ -2,6 +2,7 @@
 using Microservices.CohortExtractor.Audit;
 using Smi.Common.Messages.Extraction;
 using System.Collections.Generic;
+using System.Text.RegularExpressions;
 
 namespace Microservices.CohortExtractor.Execution.RequestFulfillers
 {
@@ -20,5 +21,12 @@ namespace Microservices.CohortExtractor.Execution.RequestFulfillers
         /// Controls what records that are fetched back should be reported as non extractable (including the reason why)
         /// </summary>
         IRejector Rejector { get; set; }
+
+        /// <summary>
+        /// Controls how modalities are matched to Catalogues.  Must contain a single capture group which
+        /// returns a modality code (e.g. CT) when applies to a Catalogue name.  E.g. ^([A-Z]+)_.*$ would result
+        /// in Modalities being routed based on the start of the table name e.g. CT => CT_MyTable and MR=> MR_MyTable
+        /// </summary>
+        Regex ModalityRoutingRegex { get; set; }
     }
 }