+++ tests/microservices/Microservices.CohortExtractor.Tests/ExtractionSuperTableTests.cs
@@ -2,25 +2,20 @@
 using System.Collections.Generic;
 using System.Data;
 using System.Data.Common;
-using System.IO;
 using System.Linq;
-using System.Text;
+using System.Text.RegularExpressions;
 using BadMedicine;
 using BadMedicine.Dicom;
 using Dicom;
 using FAnsi.Discovery;
-using FAnsi.Discovery.QuerySyntax.Update;
 using Microservices.CohortExtractor.Audit;
 using Microservices.CohortExtractor.Execution;
 using Microservices.CohortExtractor.Execution.RequestFulfillers;
-using Microservices.CohortExtractor.Execution.RequestFulfillers.Epcc;
+using Microservices.CohortExtractor.Execution.RequestFulfillers.Dynamic;
 using NUnit.Framework;
-using NUnit.Framework.Internal;
-using Rdmp.Core.Curation.Data;
-using Rdmp.Core.Curation.Data.Spontaneous;
 using Rdmp.Core.DataLoad.Triggers;
-using Rdmp.Core.Repositories;
 using Smi.Common.Messages.Extraction;
+using Smi.Common.Options;
 using Smi.Common.Tests;
 using Tests.Common;
 using TypeGuesser;
@@ -89,9 +84,11 @@ namespace Microservices.CohortExtractor.Tests
         }
 
 
-        [TestCase(DatabaseType.MicrosoftSQLServer), RequiresRelationalDb(DatabaseType.MicrosoftSQLServer)]
-        [TestCase(DatabaseType.MySql), RequiresRelationalDb(DatabaseType.MySql)]
-        public void Test_OnlyExtractableImages(DatabaseType dbType)
+        [TestCase(DatabaseType.MicrosoftSQLServer,true), RequiresRelationalDb(DatabaseType.MicrosoftSQLServer)]
+        [TestCase(DatabaseType.MySql,true), RequiresRelationalDb(DatabaseType.MySql)]
+        [TestCase(DatabaseType.MicrosoftSQLServer,false), RequiresRelationalDb(DatabaseType.MicrosoftSQLServer)]
+        [TestCase(DatabaseType.MySql,false), RequiresRelationalDb(DatabaseType.MySql)]
+        public void Test_OnlyExtractableImages(DatabaseType dbType,bool useDynamic)
         {
             var db = GetCleanedServer(dbType);
 
@@ -119,8 +116,8 @@ namespace Microservices.CohortExtractor.Tests
             
             //The strategy pattern implementation that goes to the database but also considers reason
             var fulfiller = new FromCataloguesExtractionRequestFulfiller(new[] {cata});
-            fulfiller.Rejector = new TestRejector();
-
+            fulfiller.Rejector = useDynamic? (IRejector) new DynamicRejector():new TestRejector();
+            
             foreach (ExtractImageCollection msgOut in fulfiller.GetAllMatchingFiles(msgIn, new NullAuditExtractions()))
             {
                 matches += msgOut.Accepted.Count();
@@ -193,7 +190,13 @@ namespace Microservices.CohortExtractor.Tests
             int matches = 0;
             
             //The strategy pattern implementation that goes to the database but also considers reason
-            var fulfiller = new EpccExtractionRequestFulfiller(new[] {cataCT,cataMR});
+
+            
+            var fulfiller = new FromCataloguesExtractionRequestFulfiller(new[] {cataCT,cataMR})
+            {
+                //Give it the default modality pattern (table prefix)
+                ModalityRoutingRegex = new Regex(new CohortExtractorOptions().ModalityRoutingRegex)
+            };
             
             foreach (ExtractImageCollection msgOut in fulfiller.GetAllMatchingFiles(msgIn, new NullAuditExtractions()))
             {